Create common grid detect then XRC plan
