+++ src/mx_bluesky/hyperion/experiment_plans/pin_centre_then_xray_centre_plan.py
@@ -9,8 +9,17 @@ from dodal.devices.eiger import EigerDetector
 from dodal.devices.oav.oav_parameters import OAVParameters
 
 from mx_bluesky.common.device_setup_plans.manipulate_sample import move_phi_chi_omega
+from mx_bluesky.common.device_setup_plans.utils import (
+    start_preparing_data_collection_then_do_plan,
+)
 from mx_bluesky.common.experiment_plans.change_aperture_then_move_plan import (
     change_aperture_then_move_to_xtal,
+)
+from mx_bluesky.common.experiment_plans.common_grid_detect_then_xray_centre_plan import (
+    detect_grid_and_do_gridscan,
+)
+from mx_bluesky.common.external_interaction.callbacks.common.grid_detection_callback import (
+    GridDetectionCallback,
 )
 from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback import (
     ispyb_activation_wrapper,
@@ -19,12 +28,12 @@ from mx_bluesky.common.parameters.constants import OavConstants
 from mx_bluesky.common.utils.context import device_composite_from_context
 from mx_bluesky.common.utils.log import LOGGER
 from mx_bluesky.common.xrc_result import XRayCentreEventHandler
-from mx_bluesky.hyperion.device_setup_plans.utils import (
-    start_preparing_data_collection_then_do_plan,
-)
 from mx_bluesky.hyperion.experiment_plans.grid_detect_then_xray_centre_plan import (
-    GridDetectThenXRayCentreComposite,
-    detect_grid_and_do_gridscan,
+    create_hyperion_xrc_composite,
+    create_parameters_for_flyscan_xray_centre,
+)
+from mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan import (
+    construct_hyperion_specific_features,
 )
 from mx_bluesky.hyperion.experiment_plans.oav_snapshot_plan import (
     setup_beamline_for_OAV,
@@ -34,17 +43,24 @@ from mx_bluesky.hyperion.experiment_plans.pin_tip_centring_plan import (
     pin_tip_centre_plan,
 )
 from mx_bluesky.hyperion.parameters.constants import CONST
+from mx_bluesky.hyperion.parameters.device_composites import (
+    HyperionGridDetectThenXRayCentreComposite,
+)
 from mx_bluesky.hyperion.parameters.gridscan import (
     GridScanWithEdgeDetect,
     PinTipCentreThenXrayCentre,
 )
 
 
-def create_devices(context: BlueskyContext) -> GridDetectThenXRayCentreComposite:
+def create_devices(
+    context: BlueskyContext,
+) -> HyperionGridDetectThenXRayCentreComposite:
     """
-    GridDetectThenXRayCentreComposite contains all the devices we need, reuse that.
+    HyperionGridDetectThenXRayCentreComposite contains all the devices we need, reuse that.
     """
-    return device_composite_from_context(context, GridDetectThenXRayCentreComposite)
+    return device_composite_from_context(
+        context, HyperionGridDetectThenXRayCentreComposite
+    )
 
 
 def create_parameters_for_grid_detection(
@@ -60,7 +76,7 @@ def create_parameters_for_grid_detection(
 
 
 def pin_centre_then_flyscan_plan(
-    composite: GridDetectThenXRayCentreComposite,
+    composite: HyperionGridDetectThenXRayCentreComposite,
     parameters: PinTipCentreThenXrayCentre,
     oav_config_file: str = OavConstants.OAV_CONFIG_JSON,
 ):
@@ -92,20 +108,31 @@ def pin_centre_then_flyscan_plan(
         )
 
         grid_detect_params = create_parameters_for_grid_detection(parameters)
-
         oav_params = OAVParameters("xrayCentring", oav_config_file)
+        grid_params_callback = GridDetectionCallback()
+        params = create_parameters_for_flyscan_xray_centre(
+            parameters, grid_params_callback.get_grid_parameters()
+        )
+        xrc_composite = create_hyperion_xrc_composite(composite)
+        xrc_params = create_parameters_for_flyscan_xray_centre(
+            parameters, grid_params_callback.get_grid_parameters()
+        )
+        beamline_specific = construct_hyperion_specific_features(xrc_composite, params)
 
         yield from detect_grid_and_do_gridscan(
-            composite,
-            grid_detect_params,
-            oav_params,
+            composite=composite,
+            parameters=grid_detect_params,
+            oav_params=oav_params,
+            xrc_composite=xrc_composite,
+            xrc_params=xrc_params,
+            beamline_specific=beamline_specific,
         )
 
     yield from ispyb_activation_wrapper(_pin_centre_then_flyscan_plan(), parameters)
 
 
 def pin_tip_centre_then_xray_centre(
-    composite: GridDetectThenXRayCentreComposite,
+    composite: HyperionGridDetectThenXRayCentreComposite,
     parameters: PinTipCentreThenXrayCentre,
     oav_config_file: str = OavConstants.OAV_CONFIG_JSON,
 ) -> MsgGenerator:
