+++ src/mx_bluesky/hyperion/experiment_plans/rotation_scan_plan.py
@@ -36,6 +36,9 @@ from mx_bluesky.common.device_setup_plans.manipulate_sample import (
     move_x_y_z,
     setup_sample_environment,
 )
+from mx_bluesky.common.device_setup_plans.utils import (
+    start_preparing_data_collection_then_do_plan,
+)
 from mx_bluesky.common.experiment_plans.read_hardware import (
     read_hardware_for_zocalo,
     standard_read_hardware_during_collection,
@@ -51,9 +54,6 @@ from mx_bluesky.hyperion.device_setup_plans.setup_zebra import (
     arm_zebra,
     setup_zebra_for_rotation,
     tidy_up_zebra_after_rotation_scan,
-)
-from mx_bluesky.hyperion.device_setup_plans.utils import (
-    start_preparing_data_collection_then_do_plan,
 )
 from mx_bluesky.hyperion.experiment_plans.oav_snapshot_plan import (
     OavSnapshotComposite,
