+++ tests/unit_tests/hyperion/device_setup_plans/test_utils.py
@@ -7,7 +7,7 @@ from dodal.beamlines import i03
 from ophyd.status import Status
 from ophyd_async.testing import get_mock_put
 
-from mx_bluesky.hyperion.device_setup_plans.utils import (
+from mx_bluesky.common.device_setup_plans.utils import (
     start_preparing_data_collection_then_do_plan,
 )
